+++ web/components/templates/pricing/pricingPage.tsx
@@ -13,8 +13,9 @@ import {
   TableCellsIcon,
 } from "@heroicons/react/24/solid";
 import { Disclosure } from "@headlessui/react";
-import RequestLogTable, { HELICONE_LOG_PRICING } from "./requestLogTable";
+import RequestLogTable from "./requestLogTable";
 import FeatureTable from "./featureTable";
+import { handleLogCostCalculation } from "../../../utlis/LogCostCalculation";
 
 const Slider = ({
   min,
@@ -104,28 +105,6 @@ export default function Example() {
 
   const handleRequestLogChange = (newValue: any) => {
     setRequestLogs(newValue);
-  };
-
-  const handleLogCostCalculation = (currentLogValue: number) => {
-    // calculate the estimated cost for the `currentValue` using tax brackets
-    const calculateCost = (currentValue: number) => {
-      let cost = 0;
-      let remainingValue = currentValue;
-      for (const pricing of HELICONE_LOG_PRICING) {
-        const logCount = Math.min(
-          pricing.upper - pricing.lower,
-          remainingValue
-        );
-        cost += logCount * pricing.rate;
-        remainingValue -= logCount;
-        if (remainingValue <= 0) {
-          break;
-        }
-      }
-      return cost;
-    };
-
-    return calculateCost(currentLogValue);
   };
 
   const renderLogCost = () => {
